requirements.txt
asgiref==3.2.10
asn1crypto==0.24.0
certifi==2018.1.18
chardet==3.0.4
cryptography==2.1.4
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
httplib2==0.9.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
psutil==5.4.2
psycopg2==2.8.6
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
python-dateutil==2.6.1
python-debian==0.1.32
python-debianbts==2.7.2
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
sqlparse==0.3.1
systemd-python==234
urllib3==1.22
wadllib==1.3.2
whitenoise==5.2.0
xcffib==0.5.1